requirements.txt
python-docx
lxml
pydantic
phonenumbers
tldextract
rapidfuzz
pyyaml
jsonpatch
ollama
email-validator
requests
openai